Script to print attachments (with extra confirmation)
I sometimes want to print PDF attachments from mutt.  I have a mailcap
that specifies print=evince %s when DISPLAY is set, and falls back to
this script when DISPLAY is not set.

I use evince as a sort-of print preview, and I don't want it to start
printing unexpectedly without that extra confirmation step when DISPLAY
happens to be not set.