watch-gha
I had a git watch-gha alias, but it won't let me use --help (it shows
the alias definition instead of passing --help to the external program).

I need --help because watch_gha_runs doesn't accept --wait as a short
form for whatever option it has that waits for a new build to start.